Compare Go versions numerically
This is a backport of https://github.com/openshift/release/pull/713

Original commit message:

Currently the `setup_env` function compares the required and available
Go versions alphabetically.  This doesn't work correctly when comparing
1.9 and 1.10. As a result it isn't possible to build with Go 1.10. To
avoid that issue this patch changes the function so that it compares the
versions numerically, using the `--version-sort` option of the `sort`
command.


Former-commit-id: b859a1772df8977dc5b1441dc3dfd6dc2ee71570