Avoid `logging before flag.Parse`
That message is printed before every log message because `glog` thinks
that the command line hasn't been parsed:

  ERROR: logging before flag.Parse: ...

The command line has actually been parsed, but by the `pflag` package,
not by the `flag` package. To avoid that, this patch changes the main
program so that it parses a dummy empty command line, which makes `glog`
happy.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>

Former-commit-id: d6f3c88cccea70c14b151f9163d267224aeb2acc