Add support for multiple configuration files
Currently the configuration of the server is loaded from one single
file. This patch adds support for using multiple files and directories.
For example, the AWX configuration can be stored in a `awx.yml` file and
the rules can be stored in separate files inside the `rules.d`
directory:

```
autoheal server --config-file=awx.yml --config-file=rules.d
```

This is very convenient for situations where different parts of the
configuration are owned by different persons or teams, as they can be
stored in different places and with different permissions.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>

Former-commit-id: 89ecd8d7650ff47116882f1a43da36712265b190