Make secrets optional
Currently the auto-heal server needs to read the AWX crendentials and CA
certificates from Kubernetes secret. This complicates things in
development environments, because the secrets have to be created before
using the server. To simplify things, this patch adds support for
reading the credentials and CA certificates directly from the
configuration file. For example:

```yaml
awx:
  address: https://my-awx.example.com/api
  credentials:
    username: autoheal
    password: ...
  tls:
    caFile: my-ca.pem
```

Using the `credentialsRef` and `tlsRef` sections to reference Kubernetes
is still supported and the recommended configuration when running the
server in production environments.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>

Former-commit-id: bf3e99524187b7046781e8fa88cb518996d9831c