Add `CAFile` method to AWX connection builder
This patch adds a new `CAFile` method to the builder of AWX connection
objects. This method receives the name of a file containing the CA
certificates. This is very convenient when using the AWX client in
command line programs, as most probably the CA certificates required
will be already stored in PEM files.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>

Former-commit-id: c5755fd904f5ba39938e9d32364e5097182904f6