Add script to update generated deep copy code
This patch adds a script that updates the deep copy generated code. To
use it run this command:

  hack/update-codegen.sh

And then commit to the repository the modified files.

Note that in order for these script to work correctly the definition of
the types have to follow certain conventions. For example, there must be
an unversioned definition of the types used internally by the program,
and a set of versioned definitions used externally. This patch also
adapts the poject to those conventions.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>

Former-commit-id: a62ebf400e9abe10bc1f68c574a5dc02ae87f79b