Improve consistency of log messages
This patch changes some of the log messages so that the general style is
more consistent. In general a log message should be a gramatically
correct sentence, starting with upper case and ending without
punctuation. If values, like file names, are included then the should be
enclosed in single quotes, so that it is easy to spot unusual values
like empty or blank strings:

  glog.Infof("Loading configuration from file '%s'", file)

When an error needs to be reported in the log message it should be
included at the end of the message, after a colon:

  glog.Errorf("Can't load configuration file '%s': %s", file, err)

Signed-off-by: Juan Hernandez <jhernand@redhat.com>

Former-commit-id: 5ba43c8af00bffec5304baeb8c96fa44e49743af